RaResponder: add support for NS/NA for tests
Add support to respond to neighbor solicitations in addition to router
solicitations. This is useful for tests that need to send IPv6 traffic
through a tap interface. I have previously experimented with the target
link-layer address option in the RA, but it did not seem to prevent a
subsequent NS.

As a TODO, this class should be renamed to NdResponder (and possibly
rewritten in Kotlin).

Test: test only change
Bug: 235559605
CRs-fixed: 3244429
Change-Id: I33bac507082fa89d1a5ae02cbe671ce7f5f09b07
(cherry picked from commit 42ec9d7dc23b552c42a09c20c3d17353360cb731)